No more locking on write path
This patch replaces the previous mutex-based design where the readers
took turns merging their local histograms into the shared one on a phase
change with one where the readers instead send their histograms to the
reader over a channel. The reader is then responsible for merging them
together. This does require readers to allocate on phase shifts where
they did not previously, but significantly reduces the amount of serial
work required, and should lower tail latencies by a fair amount.

Note that this design still requires writers to perform at least one
write to unblock a reader refresh (or the reader has to use a timeout).