Fix incorrect test simplification.
In
https://github.com/HdrHistogram/HdrHistogram_rust/pull/125#discussion_r1390185430
a simplification was made without discussion:

```diff
-     h.iter_recorded().collect::<Vec<_>>().len()
+     h.iter_recorded().count()
```

This is incorrect as what this test actually tests was that the iterator
has the correct length. In a perfect world both should be identical but
the PR fix a case where this was not the case as some items would be
skipped over.

In this I revert the change, add a comment and a second (redundant)
check to make sure this is not re-simplified later.